requirements.txt
Django==1.9.2
Pillow==3.1.1
SQLAlchemy==1.0.12
boto==2.39.0
click==6.3
dj-database-url==0.4.0
django-postgrespool==0.3.0
django-pylibmc==0.6.1
django-rq==0.9.0
django-storages==1.4
django-taggit==0.18.0
gunicorn==19.4.5
httplib2==0.9.2
oauthlib==1.0.3
opbeat==3.2.2
psycopg2==2.6.1
pylibmc==1.5.0
python-dateutil==2.4.2
python-instagram==1.3.2
python-twitter==2.2
pytz==2015.7
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
rq==0.5.6
simplejson==3.8.2
six==1.10.0
